Fix visual display issue
The position of the logical volume for each nested tube was set to the
position of detected photons in the sensitive detector, which looked
strange in the visualiser. Fixed the issue by adding a new hit_position
variable to the hit collection and using that instead of the logical
volume position.