Add nested tube array geometry (optical fibers) (#139)
* Add fibre info to ratpac data structure

This will allow us to pass on fiber info like angle and length to processors.

* Add nested tube arrays to docs

* Add nested tube array factory to geoBuilder

Also added a validation macro and geometry for the nested tubes.

* Add clang-formatting to new files

* Correct documentation

Changed "optical photon" back to "opticalphoton" to reflect the name of the G4 class it is referring to.

* Correct clang-format

Formatting was flagged, despite using clang-format on all edited files (I made sure the .clang-format file was being used). I've changed the flagged files to have the same format as similar existing files.

* Add sensitive detector capability for nested tubes

One can now set the cores of nested tubes as sensitive detectors. Useful if using as optical fibres.

* Add to Rat DS so fiber hit info can be saved

* Update docs for nested tube arrays

* Update nested tube documentation

* Change fiber SD to record all photons

* Remove drawing of fiber hits

* Improve efficiency of nested tube arrays

Previously, the nested tube array code would save the information for
all nested tubes in the array to the rat DS. This could potentially
consume a lot of memory. The behaviour has now been changed so that
information is only stored for nested tubes with sensitive detector
hits.

* Record attenuation hits only in fiber sensitive detectors

* Remove faulty functions from header

* Fix visual display issue

The position of the logical volume for each nested tube was set to the
position of detected photons in the sensitive detector, which looked
strange in the visualiser. Fixed the issue by adding a new hit_position
variable to the hit collection and using that instead of the logical
volume position.

* Update header

Forgot to update the header during last bugfix

* Implement clang formatting

* Add example WLS materials and use in validation

* Add nested tube hit info to ntuple output

* Fix formatting

* Fix Run.hh

A duplicate ClassDef instance was added when the main branch was merged
in.

* Remove unused ntuple variables

---------

Co-authored-by: Wilf Shorrock <w.shorrock@sussex.ac.uk>